refactor: group utility components
