+++ apps/www/src/app/components/dialog/share-link-dialog.tsx
@@ -16,7 +16,7 @@ import {
 import { Input } from "@umamin/ui/components/input";
 import { Label } from "@umamin/ui/components/label";
 import React from "react";
-import { Icons } from "../icons";
+import { Icons } from "../utilities/icons";
 
 export function ShareLinkDialog() {
   return (
