+++ apps/www/src/app/login/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Icons } from "../components/icons";
+import { Icons } from "../components/utilities/icons";
 import { Button } from "@umamin/ui/components/button";
 import {
   Card,
